% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logregNewtonRaphson.R
\name{logregNewtonRaphson}
\alias{logregNewtonRaphson}
\title{The logregNewtonRaphson function performs the Newton-Raphson optimisation
to obtain the mle estimates for coefficients in a logistic regression model}
\usage{
logregNewtonRaphson(X, y, max.iter = 30, tol = 1e-06)
}
\arguments{
\item{X}{A model matrix without intercept}

\item{y}{A response vector of 0s and 1s}

\item{max.iter}{The maximum number of iteration}

\item{tol}{A convergence criterion}
}
\value{
coefficients: vector containing the estimates

var: variance-covariance matrix of the estimates

iterations: number of iterations required
}
\description{
The logregNewtonRaphson function performs the Newton-Raphson optimisation
to obtain the mle estimates for coefficients in a logistic regression model
}
\author{
Solichatus Zahroh \email{s.solichatuszahroh@students.uu.nl}
}
